Fix code review complaints
